Remove non-Library Manager libraries from `depends`
The `depends` field of the library.properties metadata file can be used to provide a list of library dependencies that should be installed in addition during a Library Manager installation. That is the only purpose it serves. It should not contain anything other than names of libraries that are present in the Library Manager index.

Previously, this library's `depends` field contained elements that were not Library Manager libraries. This had the following effect on installation of the library:

- Classic Arduino IDE: the library is installed and the invalid elements ignored
- Arduino IDE 2.x: installation of the library fails silently
- Arduino CLI: installation of the library fails with "Error installing Forced-BME280: No valid dependencies solution found: dependency 'stdint' is not available"